supervisor/requirements.txt
pybluez 
nxt-python